Migrate authentication to blueprints
