+++ api/main/__init__.py
@@ -22,7 +22,7 @@ from flask_login import LoginManager
 from sqlalchemy import insert
 
 login_manager = LoginManager()
-login_manager.login_view = "login"
+login_manager.login_view = "auth.login"
 db = SQLAlchemy()
 ma = Marshmallow()
 bcrypt = Bcrypt()
@@ -30,7 +30,8 @@ bcrypt = Bcrypt()
 from api.main.resources.users_resource import UserResource, UsersResource
 from api.main.resources.asset_resource import Asset, Assets
 from api.main.resources.investments_resource import UserInvestedAssets, Invest
-from api.main.auth.auth import register, login, index, profile, logout
+from api.main.blueprints.auth.auth import auth
+from api.main.blueprints.index import index
 
 
 def create_app(config_name: str):
@@ -40,7 +41,6 @@ def create_app(config_name: str):
 
     from api.main.database import Currency, Users, ETF, Stock, InvestedETFs, InvestedStocks
 
-    # api.init_app(app)
     api = Api(app)
     db.init_app(app)
     ma.init_app(app)
@@ -48,13 +48,8 @@ def create_app(config_name: str):
     login_manager.init_app(app)
 
     app.register_error_handler(404, page_not_found)
-    app.add_url_rule(
-        ENDPOINTS_CONFIG.REGISTER_ENDPOINT, "register", register, methods=["GET", "POST"]
-    )
-    app.add_url_rule(ENDPOINTS_CONFIG.LOGIN_ENDPOINT, "login", login, methods=["GET", "POST"])
-    app.add_url_rule("/logout", "logout", logout, methods=["GET", "POST"])
-    app.add_url_rule("/", "index", index, methods=["GET", "POST"])
-    app.add_url_rule("/profile", "profile", profile, methods=["GET", "POST"])
+    app.register_blueprint(auth)
+    app.register_blueprint(index)
 
     with app.app_context():
         db.metadata.drop_all(db.engine)
@@ -82,7 +77,6 @@ def create_app(config_name: str):
         Invest,
         f"{ENDPOINTS_CONFIG.INVEST_ENDPOINT}/<string:username>/<string:investment_type>",
     )
-    # api.add_resource(Register, ENDPOINTS_CONFIG.REGISTER_ENDPOINT)
 
     if not CONFIG[config_name].DB_ONLY:
         sql_model = LLM(LLMType.SQL)
