+++ api/main/templates/index.html
@@ -71,9 +71,9 @@
   <body>
     <h1>Portfolio Manager</h1>
     <p>Welcome to Portfolio Manager! This web application allows you to keep track of your various investment assets.</p>
-    <p>To get started, please <a href="{{ url_for('register') }}">register</a> or <a href="{{ url_for('login') }}">login</a>.</p>
+    <p>To get started, please <a href="{{ url_for('auth.register') }}">register</a> or <a href="{{ url_for('auth.login') }}">login</a>.</p>
     {% if current_user.is_authenticated %}
-    <a href="{{ url_for('logout')}} "> logout  {{ current_user.username }}</a>
+    <a href="{{ url_for('auth.logout')}} "> logout  {{ current_user.username }}</a>
     {% else %}
     <footer>Hi {{ current_user.username }}!</footer>
     {% endif %}
