+++ api/main/templates/profile.html
@@ -74,5 +74,6 @@
     <p>Username: {{ user.username }}</p>
     <p>Email: {{ user.email }}</p>
     <p>Password: **********</p>
+    <a href="{{ url_for('auth.logout')  }}">logout</a>
   </body>
 </html>
