+++ api/main/templates/register.html
@@ -55,7 +55,7 @@
   </head>
   <body>
     <h1>Register</h1>
-    <form method="POST" action="{{ url_for('register') }}">
+    <form method="POST" action="{{ url_for('auth.register') }}">
       {{ form.csrf_token }}
       <label for="email">Email:</label>
       {{ form.email(size=30) }}
